pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "codex-logger"
version = "0.1.0"
description = "CLI logger for Codex notify payloads"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.0",
]

[project.scripts]
codex-logger = "codex_logger.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.0.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.hatch.build.targets.wheel]
packages = ["src/codex_logger"]
